+++ tests/docs/docs/conftest.py
@@ -1,8 +1,6 @@
 # # Script abc de
 # ## Test def
 
-import sys
-
 import altair as alt
 import pandas as pd
 
@@ -17,6 +15,5 @@ chart
 
 # -
 x = 1
-sys.stderr.write("abc\n")
 # -
 x * 4