+++ tests/misc/test_main.py
@@ -2,6 +2,7 @@ from click.testing import CliRunner
 
 from pheasant import __version__
 from pheasant.main import cli
+from pheasant.renderers.jupyter.client import kernel_clients
 
 
 def test_main_version():
@@ -19,6 +20,10 @@ def test_main_convert():
 
         result = runner.invoke(cli, ["run", "example.md"])
         assert result.exit_code == 0
+        assert len(kernel_clients) == 1
+        result = runner.invoke(cli, ["run", "example.md", "--restart"])
+        assert result.exit_code == 0
+        assert len(kernel_clients) == 0
 
 
 def test_main_prompt():